feat: 나라 클릭시 해당 나라로 줌인 되지 않고 이동, 나라 클릭시 해당 나라 색 변경
